app/requirements.txt
# Requirements for the Computer Vision Project
# Web framework
fastapi
uvicorn

# Vision & ML
opencv-python-headless==4.8.1.78  # stable version without GUI
numpy==1.24.4          # compatible with PyTorch & OpenCV

# Databases
pymongo
psycopg2-binary         # PostgreSQL

# Templates & Streaming
jinja2
python-multipart        # required by FastAPI for file uploads
zipfile36

# Utils
requests
